% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_user_hierarchy_structure}
\alias{connect_describe_user_hierarchy_structure}
\title{Returns a HiearchyGroupStructure object, which contains data about the
levels in the agent hierarchy}
\usage{
connect_describe_user_hierarchy_structure(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier for your Amazon Connect instance. To find the ID of your
instance, open the AWS console and select Amazon Connect. Select the
alias of the instance in the Instance alias column. The instance ID is
displayed in the Overview section of your instance settings. For
example, the instance ID is the set of characters at the end of the
instance ARN, after instance/, such as
10a4c4eb-f57e-4d4c-b602-bf39176ced07.}
}
\description{
Returns a \code{HiearchyGroupStructure} object, which contains data about the
levels in the agent hierarchy.
}
\section{Request syntax}{
\preformatted{svc$describe_user_hierarchy_structure(
  InstanceId = "string"
)
}
}

\keyword{internal}
