% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_adm_channel}
\alias{pinpoint_delete_adm_channel}
\title{Delete an ADM channel}
\usage{
pinpoint_delete_adm_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Delete an ADM channel.
}
\section{Request syntax}{
\preformatted{pinpoint$delete_adm_channel(
  ApplicationId = "string"
)
}
}

\keyword{internal}
