% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_app}
\alias{pinpoint_get_app}
\title{Returns information about an app}
\usage{
pinpoint_get_app(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Returns information about an app.
}
\section{Request syntax}{
\preformatted{svc$get_app(
  ApplicationId = "string"
)
}
}

\keyword{internal}
