% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_campaign_activities}
\alias{pinpoint_get_campaign_activities}
\title{Returns information about the activity performed by a campaign}
\usage{
pinpoint_get_campaign_activities(ApplicationId, CampaignId, PageSize,
  Token)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{CampaignId}{[required] The unique ID of the campaign.}

\item{PageSize}{The number of entries you want on each page in the response.}

\item{Token}{The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns information about the activity performed by a campaign.
}
\section{Request syntax}{
\preformatted{pinpoint$get_campaign_activities(
  ApplicationId = "string",
  CampaignId = "string",
  PageSize = "string",
  Token = "string"
)
}
}

\keyword{internal}
