% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_email_channel}
\alias{pinpoint_get_email_channel}
\title{Get an email channel}
\usage{
pinpoint_get_email_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get an email channel.
}
\section{Request syntax}{
\preformatted{svc$get_email_channel(
  ApplicationId = "string"
)
}
}

\keyword{internal}
