% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_segment_versions}
\alias{pinpoint_get_segment_versions}
\title{Returns information about your segment versions}
\usage{
pinpoint_get_segment_versions(ApplicationId, PageSize, SegmentId, Token)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{PageSize}{The number of entries you want on each page in the response.}

\item{SegmentId}{[required] The unique ID of the segment.}

\item{Token}{The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns information about your segment versions.
}
\section{Request syntax}{
\preformatted{pinpoint$get_segment_versions(
  ApplicationId = "string",
  PageSize = "string",
  SegmentId = "string",
  Token = "string"
)
}
}

\keyword{internal}
