% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_tag_resource}
\alias{pinpoint_tag_resource}
\title{Tag resource}
\usage{
pinpoint_tag_resource(ResourceArn, TagsModel)
}
\arguments{
\item{ResourceArn}{[required]}

\item{TagsModel}{[required]}
}
\description{
Tag resource
}
\section{Request syntax}{
\preformatted{pinpoint$tag_resource(
  ResourceArn = "string",
  TagsModel = list(
    tags = list(
      "string"
    )
  )
)
}
}

\keyword{internal}
