% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_adm_channel}
\alias{pinpoint_update_adm_channel}
\title{Update an ADM channel}
\usage{
pinpoint_update_adm_channel(ADMChannelRequest, ApplicationId)
}
\arguments{
\item{ADMChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Update an ADM channel.
}
\section{Request syntax}{
\preformatted{svc$update_adm_channel(
  ADMChannelRequest = list(
    ClientId = "string",
    ClientSecret = "string",
    Enabled = TRUE|FALSE
  ),
  ApplicationId = "string"
)
}
}

\keyword{internal}
