% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_apns_voip_sandbox_channel}
\alias{pinpoint_update_apns_voip_sandbox_channel}
\title{Update an APNS VoIP sandbox channel}
\usage{
pinpoint_update_apns_voip_sandbox_channel(APNSVoipSandboxChannelRequest,
  ApplicationId)
}
\arguments{
\item{APNSVoipSandboxChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Update an APNS VoIP sandbox channel
}
\section{Request syntax}{
\preformatted{svc$update_apns_voip_sandbox_channel(
  APNSVoipSandboxChannelRequest = list(
    BundleId = "string",
    Certificate = "string",
    DefaultAuthenticationMethod = "string",
    Enabled = TRUE|FALSE,
    PrivateKey = "string",
    TeamId = "string",
    TokenKey = "string",
    TokenKeyId = "string"
  ),
  ApplicationId = "string"
)
}
}

\keyword{internal}
