% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_list_dedicated_ip_pools}
\alias{pinpointemail_list_dedicated_ip_pools}
\title{List all of the dedicated IP pools that exist in your Amazon Pinpoint
account in the current AWS Region}
\usage{
pinpointemail_list_dedicated_ip_pools(NextToken, PageSize)
}
\arguments{
\item{NextToken}{A token returned from a previous call to \code{ListDedicatedIpPools} to
indicate the position in the list of dedicated IP pools.}

\item{PageSize}{The number of results to show in a single call to
\code{ListDedicatedIpPools}. If the number of results is larger than the
number you specified in this parameter, then the response includes a
\code{NextToken} element, which you can use to obtain additional results.}
}
\description{
List all of the dedicated IP pools that exist in your Amazon Pinpoint
account in the current AWS Region.
}
\section{Request syntax}{
\preformatted{svc$list_dedicated_ip_pools(
  NextToken = "string",
  PageSize = 123
)
}
}

\keyword{internal}
