% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_put_account_dedicated_ip_warmup_attributes}
\alias{pinpointemail_put_account_dedicated_ip_warmup_attributes}
\title{Enable or disable the automatic warm-up feature for dedicated IP
addresses}
\usage{
pinpointemail_put_account_dedicated_ip_warmup_attributes(
  AutoWarmupEnabled)
}
\arguments{
\item{AutoWarmupEnabled}{Enables or disables the automatic warm-up feature for dedicated IP
addresses that are associated with your Amazon Pinpoint account in the
current AWS Region. Set to \code{true} to enable the automatic warm-up
feature, or set to \code{false} to disable it.}
}
\description{
Enable or disable the automatic warm-up feature for dedicated IP
addresses.
}
\section{Request syntax}{
\preformatted{svc$put_account_dedicated_ip_warmup_attributes(
  AutoWarmupEnabled = TRUE|FALSE
)
}
}

\keyword{internal}
